refactor: common.py を chat.py にリネーム
他のファイル（rag.py, multimedia.py, riddle.py）が機能名で命名されているため、
統一感を持たせるために common.py を chat.py にリネーム。

- service/completion/common.py → chat.py
- usecase/completion/common.py → chat.py
- 関連するすべてのインポート文を更新
- テストのモックパスも更新

既存のテストがすべてパス (12 tests OK)
