+++ llm_chat/domain/service/completion/chat.py
